// src/api/model/invoiceStatusHistoryItemDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * accounting
 * Accounting API description
 * OpenAPI spec version: 1.0
 */
import type { InvoiceStatusHistoryItemDtoStatus } from './invoiceStatusHistoryItemDtoStatus';
import type { InvoiceStatusHistoryChangeDto } from './invoiceStatusHistoryChangeDto';

export interface InvoiceStatusHistoryItemDto {
  status: InvoiceStatusHistoryItemDtoStatus;
  datetime: string;
  updated?: InvoiceStatusHistoryChangeDto[];
}